property: numeric as datatype in generic

+++ trunk/property/inc/class.sogeneric.inc.php
@@ -2357,7 +2357,7 @@
 							(
 								'name'		=> 'value',
 								'descr'		=> lang('value'),
-								'type'		=> 'int',
+								'type'		=> 'numeric',
 								'nullable'	=> false,
 								'size'		=> 4,
 								'sortable'	=> true
